fix go_test when a TestMain function is defined
This broke in Go 1.8. testing.MainStart now expects a testing.testDeps
interface as its first argument. testing.internal.testdeps.TestDeps is
a public implementation of this, used by "go test".

This may not be stable, but examples/lib:lib_external_test will break
if this changes when we update our Go dependency in the future, so
we'll catch problems like this.

Fixes #273