Support gc_goopts and gc_linkopts attributes
go_library, cgo_library, go_binary, and go_test now support the
gc_goopts attribute. This is a list of flags to pass to the Go
compiler.

go_binary and go_test also support the gc_linkopts attribute. This is
a list of flags to be passed to the Go linker.

Note that _emit_go_compile_action and _emit_go_link_action have had
small interface changes and are now private. I didn't see any uses of
these functions on GitHub, and I'd like to avoid having people depend
on them in the future.

Fixes #217