go_repository: add importpath and vcs flags (#313) (#319)
* go_repository: add importpath and vcs flags (#313)

This change allows fetch_repo to fetch repositories without attempting
to determine the repo root if the user supplies the information
themselves. This allows users to be able to fetch repositories that the
vcs package would otherwise not be able to recognize.

This change also changes the behavior of the --remote flag to match the
git_repository rule.

* Respond to jayconrod@ comments in #319.

* Add BUILD rule for fetch_repo_test.go
* Stub out vcs.RepoRootForImportPath
* Make remote and vcs flags not being set in fetch_repo a hard error.
* Verify VCS is not nil before returning it in a RepoRoot.

* Fix go_repository rule for #319.

Since the vcs field in go_repository was always set to "git", fetch_repo was
always failing if the user passed in only the importpath.