Fix path to _cgo_filter_srcs script
The script name for this rule is based on the label package and
name. Previously, if the library was in the root directory of a
repository, the package name would be "", and the script name would
start with a "/", which prevented it from building.

This fix removes the "/" if the package is "".

Fixes #382