Add an integration test targets in popular repositories.
popular_repos.bash is a new test which sets up a new WORKSPACE with a
number of new_go_repository rules pointing to golang.org/x
repositories. It tests most targets (with a few exceptions) in these
repositories. This should give both the Go rules and Gazelle more
coverage on real world code.