Gazelle no longer applies build constraints by default
The Go rules now apply build constraints to files in srcs of
go_library, cgo_library, go_test, and go_binary, so it's not usually
necessary to perform this filtering in Gazelle as well. Gazelle will
not apply constraints (including GOOS and GOARCH) by default anymore.
This means generated BUILD files will be the same for all platforms.

Build tags can still be specified manually on the command flag, with
-build_tags. This can still be done through new_go_repository.

Fixes #339