Add trimpath to compile action
This removes the excessive bazel paths from the stack traces.
It does mean that stack traces no longer have an absolute path in them, but I suspect that any tooling trying to use that from a bazel build is already in trouble as it used to get a path to a temporary softlink not the real source anyway.