Gazelle: use go_prefix to determine default package name in root directory
Gazelle determines a default package name for each directory. Usually
this is the directory base name, but semantically, it should be the
last component of the import path.

With this change, the default package name for the root directory of
the repository is now the last component of the import path in
go_prefix. This accounts for repositories checked out under names that
differ from their import paths, which is always the case inside
new_go_repository.

Fixes #406