Don’t build symlink trees in bash
Only the compile step is changed so far.
It now uses separate bazel actions per file, and uses copy rather than link.
This means that you no longer need to manage directories, or wipe out the tree each time, instead each entry is updated when it needs to be.
It is also more portable.
Incremental building should be faster, and a clean build should not be much slower if at all, except in the presence of huge libraries that take a long time to copy.