Improve error message when all cgo sources are filtered out
Previously, If none of the .go sources in a cgo_library pass build
constraints, we pass an empty source list to cgo, which prints out
usage information and exits with an error code. This is confusing and
unhelpful.

With this change, we emit an error messages when there are no
buildable .go sources. This is consistent with what go_library and "go
build" do.

Related to #399