merge go_repository rules
This is a proof of concept change to merge new_go_repository into
go_repository.
Instead it autmatically runs gazelle on repositories that do not
supply a build file.
It also adds the ability to fetch a go repository by source blob
instead, with the same automatic running of gazelle if needed.
This is fully backwards compatible, but going forward would allow
gazelle to only generate go_repository rules, and to automatically
switch to source blobs when it knows it can resolve one (rather
than using the vcs)