Gazelle select (2/x): add Package, Target, PlatformStrings
Added three structs that contain metadata for a Go package in a
directory. These will be used to generate rules in place of
go/build.Context.

The main difference between these and go/build.Context is that these
structs divide files by target instead of type, and they are capable
of list platform-specific files, imports, and flags for multiple
platforms.