Gazelle: cache lookups using vcs.RepoRootForImportPath
This change fixes a performance problem in the external
resolver.

For each import path in a repository, Gazelle determines a label for
the corresponding dependency. For external dependencies, the external
name in the label is based on the repository root URL, so Gazelle
looks this up using the vcs package. The lookup process is expensive,
since it involves HTTP requests and redirects for unqualified paths.

This change caches lookup results. It also unifies the special case
logic into this cache.

Fixes #353
Related #427