Breaking up go_repositories.bzl
Splitting it into it's three core responsibilities in preparation for each of
those repsonsabilities being rewritten. This is all towards the goal of cross
compilation.
declaring the repositories we depend on -> repositories.bzl
rules for the tools used by go_repository (gazelle and fetch_repo) ->
repository_tools
rules for selecting a toolchain -> toolchain.bzl