Move def.bzl
This is a step towards refactoring def.bzl and breaking it into it's major
functional areas.
This is basically a simple move of go/def.bzl -> go/private/def.bzl and leaving
a new def.bzl in it's place that rexports all the symbols from def.bzl that
should be public, via what will be their new homes.
This is not a fully backwards compatible change, because there are symbols in
def.bzl that we do not re-export because they should never have been part of the
public interface.