Gazelle select (7/x): parse platform constraints
Add a command line flag (-platforms), which specifies the platforms
Gazelle should generate BUILD files for. Each platform has a label
(corresponding to a config_setting) and a comma-separated list of tags
that are true on that platform.

config_settings are provided for Darwin, Linux, and Windows on amd64
by default in //go/platform. Gazelle references these platforms by
default. Users can define their own config_settings and reference them
on the command line.

The command line flag is parsed and added to generator.Generator, but
it's not used yet.

Related #409