Rule implementation functions are not general purpose
go_library_impl -> emit_library_actions

a rule implementation function is not a useful building block for other
functions, and treating it as one was causing us to leak implementation details
into our public interface.
Now the action emitting functions are decoupled from the code that looks into
attributes, we are free to change and clean up the attributes and return values
of our rules, which is neccesary in the change to true providers.