Support select in cgo_library srcs and deps
cgo_library should allow srcs and cdeps to be arbitrary expressions
that produce lists of appropriate sources. In particular, they should
allow calls to select, which allows sources and dependencies to be
included conditionally.

This change introduces a new rule, _cgo_select_files, which copies
files that match a given set of suffixes (possibly inverted).
_setup_cgo_library and related functions are rewritten to
use this. Some other small fixes are included, such as demangling cgo
outputs.

In the future, we'll use output groups with filegroups instead of
_cgo_select_files, but they aren't supported in Bazel 0.4.4 (our
current minimum).

Fixes #510