Fix go_test chdir to package directory
The main go file generated by go_test includes a call to os.Chdir to
change to the package directory so that data files may be accessed by
package-relative paths. This is the same behavior as "go test".

The chdir now includes the workspace directory, which is necessary
for tests in remote workspaces. We also detect errors when changing
directories.

Fixes #540