Full bazel tests from inside bazel
This makes a new system using bazel rules and sh_test to run bazel tests from
inside bazel.
This moves only cgo_library_root_dir as a prototype, but it should hold for all
the others.

To run the same tests as before do any of:
    bazel test --test_tag_filters=-bazel //...
To only run the "non_bazel_tests"
    bazel test --test_tag_filters=bazel //...
To run all the tests you ought to be running
    bazel test //...

Better method of running tests when trying to work fast
    bazel test --test_size_filters=-large //...

Even better methods of running tests when trying to work fast that are wrong at
the moment because we are not being disciplined about our test tagging
    bazel test --test_size_filters=small //...
    bazel test --test_timeout_filters=short //...