Gazelle: treat release tags as unknown
Gazelle will no longer filter files based on release tag
(e.g., "go1.8"). Release tags are not currently part of the
config_settings used to select files, and they are difficult to detect
at that point. So we'll include files with these tags, whether or not
they are negated, and we'll let the rules sort it out.

Fixes #563