Adding rules for running gazelle
These are still experimental, and subject to change, but they seem useful.

Add
    load("@io_bazel_rules_go//go:def.bzl", "gazelle")
    gazelle(name = "gazelle")
to your build file, and then do
    bazel run //:gazelle
to automatically fix all build files.
You can make as many of these rules as you like with different names to run
gazelle in different modes or with different options.