Go rules for new proto_library
This works by adding a go_proto_library rule that takes a proto_library as it's
input and producing the same results as go_library would do using those
generated files.
It extracts the configuration to two toolchains, one called proto that supplies
protoc and one called go_proto that adds the go plugin and controls how it is
invokoked.
To add extra sources, you can make a go_library rule that uses the embed feature
to consume the go_proto_library.