Gazelle: fix load statements before merge
This change is the first in a series that will remove cgo_library from
Gazelle, now that it's no longer needed. The new functionality here is
not wired in yet so that the large change can be reviewed in smaller
pieces.

This adds a new function, fixLoads, which modifies load statements in
existing BUILD files. Symbols will only be loaded for rules that will
actually used. Multiple load statements may be consolidated into a
single statement. Load statements and symbols not generated by Gazelle
will not be modified.

This function will be applied after cgo_library rules are replaced or
merged with go_library rules. The cgo_library symbol should not be
loaded after this.