Fix up creation of GoLibrary providers
We create them directly inside emit_library_actions now, so we dont have to do a
struct to provider field by field copy.
the cgo_object is now passed in it's own provider, this is partly to enable a
future clean up of the cgo chain, and partly because it contains mutable fields.
This leaves GoLibrary as something that can now be put in a depset.