Action generation on the toolchain
This separates the action generating code (mostly) from the rule
implementations.
Action generators are now registered with the toolchain, which means toolchains
can swap out the implementation of the actions at will.
This required a lot of breaking up of files, and to make it easy to understand
and reason about, the rule parts were put in a separate folder from the action
parts.
The only thing that should depend on things in actions/ is the toolchain
constructor, all access from rules go through the toolchain.

This should not be a functional difference, but it does enable a bunch of new
things. This also an example of what I would like the cc toolchain to look like
so we can spit out compile and link actions inside go_library.