Gazelle: initial support for generalized directive comments
Added config.ParseDirectives, which looks for comments of the form

    # gazelle:key value

Directives related to the configuration will be applied to a copy of
the current configuration, which is returned. This is passed down
through subdirectories by packages.Walk.

A full list of directives is returned. These may be processed by other
packages, for example gazelle:ignore and gazelle:exclude.