Simplify the SDK story
This makes it so that `@go_sdk` is always the active sdk.
This means that the version is now set by which sdk you map to that name, not by
constraints in the toolchain. It also means that you can find the sdk and source
by a common name without going through the active toolchain.

This is a prerequiste for compiling the stdlib on demand (see #814)