Build tools with a toolchain
This merges the implementations of go_tool_binary and go_binary
We then substitute the compile and link actions for ones that work without the
intermediate tools in the boostrap toolchain.
This means the tools are now built in the normal way, rather than with `go
build`

This is more work towards #814, as it will enable me to change the way we handle
the environment.