Gazelle: support empty prefixes
Gazelle now allows go_prefix to be empty. It must still be set either
on the command line or in a go_prefix rule. When the prefix is empty,
all imports not in the standard library are resolved
locally. Dependencies will not point to the vendor directory or to
external repositories.

* Gazelle now hard-codes a list of standard packages from the SDK that
  was used to build it.
* All imports (including standard imports) are now added to
  packages.GoTarget. The resolve package filters out standard
  libraries.

Fixes #870