Structured compilation modes
This changes the way we handle compilation modes to be a fully strucutred form.
It does not change (yet) the way we pick the mode, but it gives all the actions
a fully structured mode that is capapble of supporting all the combinations.

As a bonus we get full support for msan in this pr.