Transition to aspect based builds
We can now build libraries in multiple modes using aspects.
This allows the mode to be controlled by the leaf rules, so you can specify
things like `pure = "on"` on a go_binary and have it build using the stdlib that
does not depend on libc.
This abandons all the previous output group behaviors.
This also moves the debug and strip controls in to the build mode. In theory
this means we can build stripped and unstripped variants at will in a single
build, or build tests in debug mode but binaries in release mode, all in a
single bazel invocation.
The documentation still needs more updates, but I will do that in another cl.