Gazelle: support prefix directive
Gazelle now supports '# gazelle:prefix <prefix>' directives in build
files. Such a directive will set the Go importpath prefix for rules in
that directory and in subdirectories.

Prefixes may be empty, but rules cannot be generated in a directory
where the prefix is set to the empty string (an error will be
reported).

Fixes bazelbuild/bazel-gazelle#6