Fix go proto library for once and for all
This uses an aspect to collect up the full set of proto imports and what their
go import path should be, and then hands that to the proto compile line.
It then attempts to map the outputs to the expected outputs by basename.
All of this means that the option go_package lines are not needed for a correct
build.