Support transitive stamping
go_library rules now support the x_def attribute.
The GoSource provider carries these to the archive.
The GoArchive provider carries the x_def values from it's GoSource as well as
from the full transitive set of archives it depends on.
If an x_def entry has no package it is given the importpath of the library the
GoSource was built from.

Fixes #628