go_repository: don't pass default values on the command line
go_repository will no longer pass arguments to Gazelle on the command
line when the corresponding attributes are not explicitly set.

This gives Gazelle more control over the defaults. For example,
Gazelle allows an inferred proto mode to override the default mode,
but not when the default mode is explicitly requested.

Related bazelbuild/bazel-gazelle#104