Reverse sense of importpath and importmap
importpath is the source importpath. Packages can reference other
packages using import declarations with this string.

importmap is the actual importpath. The compiler and linker use this,
and it's visible in stack traces. This only needs to be set when
multiple packages with the same import path would otherwise be linked.

Related #1237