Warn when linking archive with the same package path
The link wrapper will now print a warning when linking more than one
package with the same package path. This is dangerous, since only one
of the packages will be linked, and the compiler and linker may see
different code.

This will be a warning in 0.11.0 and an error in releases after that.

Also:

* Corrected documentation for direct / transitive members of
  GoArchive.
* Marked searchpath and searchpaths members as deprecated. They will
  not be needed when we switch to importcfg after dropping Go 1.8.

Related #1327