Add skeleton code for build-time code analysis framework
Add a barebone implementation of a build-time code analysis framework
which code-generates a (no-op) checker binary that is run alongside
the compiler. Key features introduced in this CL:

* go_checker rule: determines what checks go into the checker binary.
* go_register_checker: a synthetic repo used to store the label of a
    user-defined go_checker target.
* generate_checker_main: a tool that code-generates the checker binary.
* Basic test cases that verify that the default checker binary is
    generated for go_library, go_test, and go_binary targets.

In forthcoming CLs, I will extend this implementation so that the
checker binary imports check libraries that are passed to the go_checker
rule.