Allow checks to be configured using JSON file
Custom go_checker checks can now be configured using a JSON
configuration file. The format of the configuration file will
be specified in documentation that will be added in forthcoming
commits.

Also, add extra_files into the runfiles for _bazel_test_script so
that changes to the contents of extra_files causes bazel_test targets
to be re-executed. (Adapted from #1524 on master.)