Add srcs attribute to GoStdLib provider
The src directory is not added to the files attribute so it needs to be
explicitly depended upon. Therefore, this change does not have any
effect on downstream rules, and does not add to the sandboxing overhead.

The src directory is useful for some linter invocations which can
depend on sandboxed stdlib instead of expecting the users to have an
installed stdlib.