nogo: pass explicit path to .x output file
The compile builder and nogo assume the .x file to be written is in a
location based on the .a file, but analysis generates an .x output
path based on the import path, and the .a output path may vary
slightly from that.

Rather than trying to fix this assumptions, this change makes the
compile builder accept an -x parameter with the output path.

A logical extension of this would be to make paths explicit for .x
files of all dependencies. However, it would be better to get rid of
.x files altogether and store the data in a new section within .a
files instead (#1847).

Fixes #1846