go_proto_library: support proxy proto_library
In some repositories, a proto_library target will be declared for each
.proto file, even for .proto files in the same package. A
proto_library target with deps and no srcs may be used to combine
several proto_library rules.

With this change, go_proto_library treats a "proxy" proto_library as
if it contained the direct srcs of its deps.

Fixes #1857