go_toolchain: include a builder binary in the toolchain
Until now, we've used go_tool_binary to build various executables
needed to perform actions, and we've depended on those directly
through the "_builders" attribute of Go rules. This tightly couples
the rules with the way actions are performed, which may be different
with other toolchains (e.g., gccgo).

This PR makes the following changes:

* A new rule, go_toolchain_binary, may be used to build simple binaries.
* The SDK repository rules now declare a "builder" target that uses
  this rule to build an executable from sources in
  @io_bazel_rules_go//go/tools/builders.
* go_toolchain rules now require a "builder" attribute, and in the SDK
  repository rules, the above target is used.
* The "builder" executable is exposed as a private field in the
  ToolchainInfo object returned by go_toolchain. This may be
  referenced by the emit functions (though it is not yet).

Updates #1891