Make googleapis work on FreeBSD.
The patches that are shipped with googleapis have the intent of removing
files from the input. Though GNU patch does this properly, the copy of
patch shipped with FreeBSD does not do this explicitly. It needs the -E
flag to do so.

This makes it possible to build https://github.com/buildbarn on FreeBSD
11.2-STABLE.