cgo: don't require cc_library to produce files
When we analyze C dependencies in cgo, we try to locate static
archives that can be packed with the Go archives the compiler
produces. Currently, we expect cc_library to produce some kind of
static archive. It doesn't actually produce anything though if all .c
files are filtered out.

With this change, if a cc_library file doesn't produce any files, it
won't be built or packed.

Fixes #1930