Add iOS as a platform
iOS is a special variant of GOOS=darwin. In order to support the new
CC toolchain constraints that come with bazel 0.26, we need to propagate
the same OS that what bazel expects, or else CGo won't work because it
can't find a toolchain.

This needs Bazel 0.26.1 as 0.26.0 has a bug regarding this [1].

1. https://github.com/bazelbuild/bazel/pull/8522

Fixes #2079

Signed-off-by: Steeve Morin <steeve@zen.ly>