Split compiled code and export data for each archive
Instead of this, we produce two separate files (both still in the ar
format): a .a file that contains compiled code, and a .link.a file that
contains export data.

The advantage of producing separate files is that when the
implementation of a function changes in a .go file, only the .a file
built from that source changes; the .link.a file stays the same.

To enable this build mode, use the following configuration flag:

    --define=go.linkobj=1

Signed-off-by: Steeve Morin <steeve@zen.ly>