go/tools/bazel_testing: don't panic on setup error
setupWorkspace returned a nil cleanup function if it encountered an
error. It's caller panicked when calling this before checking for an
error. Now, setupWorkspace returns a trivial cleanup function instead.

Also, cleanup now returns an error. If cleanup returns an error and
the code was 0, TestMain sets the exit code to 1.

Updates #2233