Fix minor bugs in bazel_tests
Create a query:isolate entry in the config so that bazel_test rules
can be written that call bazel query.

Excitingly, --output_base is not the outermost directory, so it ends
up with some files left in $HOME. Change to --output_user_root so that
everything goes into the temp dir.